LayoutTests/platform/mac/accessibility/aria-toolbar.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../../../fast/js/resources/js-test-style.css">
<script>
var successfullyParsed = false;
</script>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<div role="toolbar" aria-label="test toolbar">
<li>item one
<li>item tw
</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that the aria-toolbar role correctly returns a toolbar.");

    if (window.accessibilityController) {

          var body = document.getElementById("body");
          body.focus();
          var obj = accessibilityController.focusedElement.childAtIndex(0);

          var succeeded = obj.role == "AXRole: AXToolbar";
          shouldBe("succeeded", "true");
    }

    successfullyParsed = true;
</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
